Replace `bash-parser` by `(at)ericcornelissen/bash-parser`
Remove the package `bash-parser` [1] by a fork by me of the same name
but published in a namespace [2]. The motivation for this is that the
original is unmaintained and this fork comes with dependency maintenance
- thus far in the form of the removal of non-OSI packages.

Do note that this fork does not promise to provide anything beyond
dependency maintenance. A proper replacement of `bash-parser` should be
preferred if available.

--
1. https://www.npmjs.com/package/bash-parser
2. https://www.npmjs.com/package/@ericcornelissen/bash-parser