Deduplicate workspace names
It is possible that the manifest workspaces include the root workspace.
In this case `ConfigurationChie#getAvailableWorkspaceName()` would
return the root name twice. This in turn would result in the root
package being included twice and throw a `ConfigurationError` because of
a duplicate package name.

To address this we deduplicate the workspace names.

Fixes #390