fix: simplify husky plugin
The one change needed to support Husky v9 is to stop follow the configured git hook path and always check in `.husky`, the rest of the logic in #500 causes issues rather than helps.

I simply reverted the plugin to the implementation prior to #500 and simply added the `false` to `getGitHookPaths('.husky', false)` and that makes all tests pass – both Husky v8 and Husky v9 – and works in both kinds of projects for me.

Side note: Husky v9 is backwards compatible with Husky v8 and says that one shouldn't have to do anything when upgrading for it to work.