feat(plugins): add plugin for xo config
Adds a plugin for [xo](https://github.com/xojs/xo#config). Essentially its an opinionated wrapper
around eslint with a bunch of bundled rules and plugins, and the config object is just the eslint
config object with some extra options. So to keep it consistent with the eslint plugin and avoid
duplication I've just used the eslint helper to parse out any additional eslint plugins that have
been added in the xo config.