Fix scalablity of knip when processing a large number of files
On a large number of files, knip is crashing with out-of-memory. This is caused by the fact
that all of the source file text and programs are being held for the entirety of the processing.

There is an explicit factory.deletePrincipal being done after every principal is processed,
in the `!isIsolateWorkspaces && isSkipLibs && !isWatch` case, but the `principals` array
is still holding onto the principal, so the memory is not released.

As we iterate through the principals, I clear the entry if it is deleted. I chose to set
the principals entry to `undefined`, rather than manipulating the list, for code simplicity.