Fix iterating over namespaces imported using '*' (star) reporting unused classes
Using ``import * as classes from...``, and then doing an iteration ``for (... in classes)``
erroneously marks the imported classes as 'not used'.

The import-star path leads to a namespace being added, and the reference to the namespace
due to iteration was not being considered, although Object.keys-type accesses were.

While fixing the ...in case, I realized the ...of iteration case was also potentially
broken, although it was very difficult to build a test for that particular edge case ...
it is highly dependent on the tsconfig options.