Use codeclimate-types
This replaces the custom CodeClimate types with the `codeclimate-types`
package. Reusing existing types is less error prone that maintaining
them internally.

This change uncovered that end positions were missing. This doesn’t
affect GitLab, which doesn’t use this information (yet), but it might
affect other tooling that expects it.