Avoid position to calculate CodeClimate fingerprint
CodeClimate fingerprints should uniquely identify the issue. The problem
with using positional information is that though it adds to the
uniqueness, it significantly reduces the identification part.

To clarify, let’s say we have the following file:

```js
// file.js
export function used() {}

export function unused() {}
```

The function `unused` is detected as unused by knip. The hash is
calculated from:
- `'/runner/file.js'` (filePath)
- `'unused'` (message)
- `54` (pos)

Let’s say this results in the hash `abc`.

This hash has two problematic factors:
1. It depends on an absolute path, which is system dependant. For
   example a GitLab users may switch to a different runner or Docker
   image.
2. If any characters are inserted or deleted before `unused`, the
   fingerprint changes.

Let’s say the user now makes a change:

```diff
  // file.js
  export function used() {}
+ // Some inserted text

  export function unused() {}
```

The new hash is calculated from:
- `'/runner/file.js'` (filePath)
- `'unused'` (message)
- `76` (pos)

Let’s say this results in the hash `def`.

Now the GitLab merge request  UI will tell the user that the first
problem with `hash` is resolved. Also a new problem is introduced with a
hash `def`.

By omitting the positional information, the same hash would be
preserved, and the GitLab UI will tell the user the merge request didn’t
affect the results.

I also added an implementation to ensure hashes are always unique in a
single run. This is based on
[eslint-formatter-gitlab#48](https://gitlab.com/remcohaszing/eslint-formatter-gitlab/-/merge_requests/48).

No tests were modified, because the hashes are explicitly omitted from
testing.