Add missing dependency on remark-gfm
This dependency is used as a transient dependency of
`@astrojs/markdown-remark`. This is configured correctly in
`tsconfig.json` so that the MDX language server and TypeScript plugin
understand it. Previously this dependency was deduped, which worked
fine. However, with the way pnpm installs dependencies, transient
dependencies aren’t available directly. This causes a crash in the
editor integrations.

To solve this, `remark-gfm` was added as a direct dependency.